@@ -1,9 +1,12 @@
 import {Pipe, PipeTransform} from '@angular/core';
 
-@Pipe({ name: 'myRoundCoords' })
+@Pipe({ 
+    name: 'myRoundCoords',
+    standalone: true,
+ })
 
 export class RoundCoords implements PipeTransform {
-    transform(value: number, args: string[]) {
+    transform(value: number) {
         if (value) {
             if (Math.abs(value ) < 360) {
                 return Math.round(value * 10e5) / 10e5;
@@ -11,8 +14,9 @@ export class RoundCoords implements PipeTransform {
                return Math.round(value * 10e1) / 10e1 ;
             }
 
+           
         }
-
+        return value;
 
     }
 }
